Pruebas de las tablas con enrutado - para Julian
